Fix sync function
