+++ src/js/components/core/SideBar/SideNavBar.js
@@ -1,4 +1,3 @@
-
 const api = window.ModuleApi;
 const React = api.React;
 const CoreActions = require('../../../actions/CoreActions.js');
@@ -11,9 +10,8 @@ const RB = api.ReactBootstrap;
 const {Glyphicon} = RB;
 const Image = require('react-bootstrap/lib/Image.js');
 const style = require("./Style");
-const git = require('../GitApi.js');
 const gogs = require('../login/GogsApi.js');
-const pathFinder = require('path');
+const sync = require('./GitSync.js');
 class SideNavBar extends React.Component{
   handleCreateProject(){
     CoreActions.showCreateProject("Languages");
@@ -24,46 +22,7 @@ class SideNavBar extends React.Component{
   }
 
   handleSyncProject(){
-    const path = api.getDataFromCommon('saveLocation');
-    const user = CoreStore.getLoggedInUser();
-    if (user) {
-      git(path).save('Updating with Door43', path, function() {
-          var manifest = api.getDataFromCommon('tcManifest');
-          if (manifest.repo) {
-            var urlArray = manifest.repo.split('.');
-            urlArray.pop();
-            var finalPath = urlArray.pop().split('/');
-            var repoPath = finalPath[1] + '/' + finalPath[2];
-            var remote = 'https://' + user.token + '@git.door43.org/' + repoPath + '.git';
-            git(path).update(remote, 'master', false, function(err){
-              if (err) {
-                var Confirm = {
-                  title: 'You don\'t have permission to push to this repository.',
-                  content: "Would you like to create a new Door43 project?",
-                  leftButtonText: "No",
-                  rightButtonText: "Yes"
-                }
-                api.createAlert(Confirm, function(result){
-                  if(result == 'Yes') {
-                    const projectName = path.split(pathFinder.sep);
-                    gogs(user.token).createRepo(user, projectName.pop()).then(function(repo) {
-                      var newRemote = 'https://' + user.token + '@git.door43.org/' + repo.full_name + '.git';
-                      git(path).update(newRemote, 'master', true, function(){});
-                    });
-                  }
-                });
-              } else {
-                alert('Update succesful');
-              }
-            });
-          } else {
-                alert('There is no associated repository with this project. Would you like to make one?');
-          }
-      });
-    } else {
-      alert('Login then try again');
-      CoreActions.updateLoginModal(true);
-    }
+    sync();
   }
 
   handleReport(){
